requirements.txt
# Core SDR Library
pyrtlsdr>=0.2.93

# Data Processing
numpy>=1.24.0
scipy>=1.10.0

# GPS Support
pyserial>=3.5
pynmea2>=1.18.0

# Configuration
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
simplekml>=1.3.6

# Data Storage
pandas>=2.0.0

# Utilities
tqdm>=4.65.0  # Progress bars
colorama>=0.4.6  # Colored terminal output
